+++ lib/metro/stubs/model_stub.dart
@@ -2,8 +2,6 @@
 String modelStub({String? modelName}) => '''
 import 'package:nylo_framework/nylo_framework.dart';
 
-/// $modelName Model.
-
 class $modelName extends Model {
   $modelName();